adds some more stuff to helper module, updates headcount and enrollment repo
