+++ lib/headcount_analyst.rb
@@ -18,21 +18,6 @@ class HeadcountAnalyst
   }}
   end
 
-  def load_data(args)
-    args[:enrollment].keys.each do |key|
-      if key == :high_school_graduation
-        @hs_key = CSV.open(args[:enrollment][key], headers: true, header_converters: :symbol)
-      else
-        @kg_key = CSV.open(args[:enrollment][key], headers: true, header_converters: :symbol)
-      end
-    end
-  end
-  # def load_data(args)
-  #   @data = CSV.open(args[:enrollment][key], headers: true, header_converters: :symbol)
-  # end
-
-  def data_cleaner ;end
-
   def kindergarten_participation_rate_variation(district, comparison)
     comparison = comparison.values[0]
     district_info = year_and_rate(district)
